@@ -11,6 +11,6 @@
     "react-dom": "^19.1.0"
   },
   "engines": {
-    "node": ">=18"
+    "node": ">=22.0.0"
   }
 }
