@@ -7,7 +7,7 @@ from requests_toolbelt.multipart import decoder
 import uuid
 from datetime import datetime
 
-app = func.FunctionApp()
+app = func.FunctionApp(http_auth_level=func.AuthLevel.FUNCTION)
 
 CONNECTION_STRING = os.environ.get("CONNECTION_STRING_PRODUCT") if os.environ.get("IS_MAIN_PRODUCT") == "true" else os.environ.get("CONNECTION_STRING_TEST")
 
@@ -108,6 +108,10 @@ def create_event(req: func.HttpRequest) -> func.HttpResponse:
             for f in required_fields:
                 if not data.get(f):
                     return error_response(f"{f}は必須です")
+        # creatorが空文字やNoneの場合は必ず'0738'にする
+        creator = data.get("creator")
+        if not creator:
+            creator = "0738"
         with get_db_connection() as conn:
             cursor = conn.cursor()
             cursor.execute(
@@ -122,7 +126,7 @@ def create_event(req: func.HttpRequest) -> func.HttpResponse:
                 to_db_date(data.get("deadline")),
                 data.get("location"),
                 int(data.get("max_participants")) if data.get("max_participants") else None,
-                str(data.get("creator", "0738")),
+                str(creator),
                 data.get("summary"),
                 data.get("detail"),
                 data.get("image"),
@@ -171,48 +175,6 @@ def get_draft(req: func.HttpRequest) -> func.HttpResponse:
         logging.error(str(e))
         return error_response("DB error", 500)
 
-@app.route(route="update_event", methods=["PUT"])
-def update_event(req: func.HttpRequest) -> func.HttpResponse:
-    event_id = int(req.route_params.get('event_id'))
-    try:
-        data, _ = parse_multipart(req)
-        with get_db_connection() as conn:
-            cursor = conn.cursor()
-            cursor.execute("SELECT creator FROM EVENTS WHERE event_id=?", (event_id,))
-            row = cursor.fetchone()
-            if not row:
-                return error_response("イベントが存在しません", 404)
-            event_creator = row.creator if hasattr(row, "creator") else row[0]
-            request_creator = str(data.get("creator", ""))
-            if not request_creator or request_creator != str(event_creator):
-                return error_response("イベント作成者のみ編集可能です", 403)
-            cursor.execute(
-                '''
-                UPDATE EVENTS SET event_title=?, event_category=?, event_datetime=?, deadline=?, location=?, max_participants=?, description=?, content=?, image=? WHERE event_id=?
-                ''',
-                data.get("title"),
-                int(data.get("category")) if data.get("category") else None,
-                to_db_date(data.get("date")),
-                to_db_date(data.get("deadline")),
-                data.get("location"),
-                int(data.get("max_participants")) if data.get("max_participants") else None,
-                data.get("summary"),
-                data.get("detail"),
-                data.get("image"),
-                event_id
-            )
-            cursor.execute("DELETE FROM EVENTS_KEYWORDS WHERE event_id=?", event_id)
-            if data.get("keywords"):
-                for kw in data["keywords"]:
-                    cursor.execute("INSERT INTO EVENTS_KEYWORDS (event_id, keyword_id) VALUES (?, ?)", event_id, int(kw))
-            conn.commit()
-        return func.HttpResponse(json.dumps({"message": "イベント更新完了", "event_id": event_id}), mimetype="application/json", status_code=200)
-    except Exception as e:
-        import traceback
-        tb = traceback.format_exc()
-        logging.error(tb)
-        return error_response(str(e), 500, tb)
-
 @app.route(route="delete_event/{event_id}", methods=["DELETE"])
 def delete_event(req: func.HttpRequest) -> func.HttpResponse:
     event_id = req.route_params.get('event_id')
@@ -248,7 +210,7 @@ def get_categories(req: func.HttpRequest) -> func.HttpResponse:
     try:
         with get_db_connection() as conn:
             cursor = conn.cursor()
-            cursor.execute("SELECT category_id, category_name FROM CATEGORYS")
+            cursor.execute("SELECT category_id, category_name FROM CATEGORIES")
             rows = cursor.fetchall()
             result = [{"category_id": row.category_id, "category_name": row.category_name} for row in rows]
         return func.HttpResponse(json.dumps(result), mimetype="application/json")
@@ -280,7 +242,7 @@ def search_events(req: func.HttpRequest) -> func.HttpResponse:
                     sql = '''
                         SELECT e.*, c.category_name
                         FROM events e
-                        LEFT JOIN CATEGORYS c ON e.event_category = c.category_id
+                        LEFT JOIN CATEGORIES c ON e.event_category = c.category_id
                     '''
                     cursor.execute(sql)
                     columns = [column[0] for column in cursor.description]
@@ -295,7 +257,7 @@ def search_events(req: func.HttpRequest) -> func.HttpResponse:
                     sql = '''
                         SELECT e.*, c.category_name
                         FROM events e
-                        LEFT JOIN CATEGORYS c ON e.event_category = c.category_id
+                        LEFT JOIN CATEGORIES c ON e.event_category = c.category_id
                         WHERE e.event_id = ?
                     '''
                     cursor.execute(sql, (event_id_int,))
@@ -341,14 +303,115 @@ def get_event_detail(req: func.HttpRequest) -> func.HttpResponse:
                 (event_id,)
             )
             row = cursor.fetchone()
+            # ...
             if row:
                 keys = ["event_id", "event_title", "event_category", "event_datetime", "deadline", "location", "max_participants", "current_participants", "creator", "description", "content", "image", "is_draft"]
                 event = dict(zip(keys, row))
+                # ...
                 return func.HttpResponse(json.dumps(event, default=str), status_code=200, mimetype="application/json")
             else:
+                # ...
                 return error_response("イベントが見つかりません", 404)
     except Exception as e:
+        # ...
         return error_response(str(e), 400)
+
+
+@app.route(route="events/{event_id}", methods=["PUT"])
+def update_event(req: func.HttpRequest) -> func.HttpResponse:
+    event_id = int(req.route_params.get('event_id'))
+    try:
+        # 作成者チェック
+        content_type = req.headers.get("Content-Type", "")
+        data = {}
+        image_path = None
+        if content_type.startswith("multipart/form-data"):
+            body = req.get_body()
+            multipart_data = decoder.MultipartDecoder(body, content_type)
+            for part in multipart_data.parts:
+                content_disposition = part.headers.get(b'Content-Disposition', b'').decode()
+                if 'filename=' in content_disposition:
+                    filename = content_disposition.split('filename="')[1].split('"')[0]
+                    ext = os.path.splitext(filename)[1]
+                    user_id = "edit"
+                    unique_id = uuid.uuid4().hex[:8]
+                    save_name = f"{datetime.now().strftime('%Y%m%d%H%M%S')}_{user_id}_{unique_id}{ext}"
+                    save_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', 'front', 'public', 'images'))
+                    os.makedirs(save_dir, exist_ok=True)
+                    save_path = os.path.join(save_dir, save_name)
+                    with open(save_path, "wb") as f:
+                        f.write(part.content)
+                    image_path = f"images/{save_name}"
+                    data["image"] = image_path
+                else:
+                    name = content_disposition.split('name="')[1].split('"')[0]
+                    value = part.text
+                    if name == "keywords":
+                        if "keywords" not in data:
+                            data["keywords"] = []
+                        data["keywords"].append(value)
+                    else:
+                        data[name] = value
+            if "image" not in data:
+                data["image"] = None
+        else:
+            data = req.get_json()
+            image_path = data.get("image")
+
+        # DBからイベント取得して作成者チェック
+        conn = get_db_connection()
+        cursor = conn.cursor()
+        cursor.execute("SELECT creator FROM EVENTS WHERE event_id=?", event_id)
+        row = cursor.fetchone()
+        if not row:
+            return func.HttpResponse(json.dumps({"error": "イベントが存在しません"}), mimetype="application/json", status_code=404)
+        event_creator = row.creator if hasattr(row, "creator") else row[0]
+        request_creator = str(data.get("creator", ""))
+        if not request_creator or request_creator != str(event_creator):
+            return func.HttpResponse(json.dumps({"error": "イベント作成者のみ編集可能です"}), mimetype="application/json", status_code=403)
+        def to_db_date(val):
+            if not val or (isinstance(val, str) and val.strip() == ""):
+                return None
+            if isinstance(val, str) and 'T' in val:
+                try:
+                    date_part, time_part = val.split('T')
+                    if len(time_part) == 5:
+                        time_part += ':00'
+                    return f"{date_part} {time_part}"
+                except Exception:
+                    return val
+            return val
+
+        with get_db_connection() as conn:
+            with conn.cursor() as cursor:
+                cursor.execute(
+                    '''
+                    UPDATE EVENTS SET event_title=?, event_category=?, event_datetime=?, deadline=?, location=?, max_participants=?, description=?, content=?, image=? WHERE event_id=?
+                    '''
+                    ,
+                    data.get("title"),
+                    int(data.get("category")) if data.get("category") else None,
+                    to_db_date(data.get("date")),
+                    to_db_date(data.get("deadline")),
+                    data.get("location"),
+                    int(data.get("max_participants")) if data.get("max_participants") else None,
+                    data.get("summary"),
+                    data.get("detail"),
+                    data.get("image"),
+                    event_id
+                )
+                # キーワード更新（全削除→再登録）
+                cursor.execute("DELETE FROM EVENTS_KEYWORDS WHERE event_id=?", event_id)
+                if data.get("keywords"):
+                    for kw in data["keywords"]:
+                        cursor.execute("INSERT INTO EVENTS_KEYWORDS (event_id, keyword_id) VALUES (?, ?)", event_id, int(kw))
+                conn.commit()
+        return func.HttpResponse(json.dumps({"message": "イベント更新完了", "event_id": event_id}), mimetype="application/json", status_code=200)
+    except Exception as e:
+        import traceback
+        tb = traceback.format_exc()
+        logging.error(tb)
+        return func.HttpResponse(json.dumps({"error": str(e), "trace": tb}), mimetype="application/json", status_code=500)
 
 
 @app.route(route="get_participants")
@@ -357,6 +420,8 @@ def get_participants(req: func.HttpRequest) -> func.HttpResponse:
     event_id = req.params.get('event_id')
     if not event_id:
         return func.HttpResponse("event_id is required", status_code=400)
+        print("[DEBUG] get_event_detail req.params:", getattr(req, 'params', None))
+        print("[DEBUG] get_event_detail req.route_params:", getattr(req, 'route_params', None))
 
     conn_str = os.environ["CONNECTION_STRING"]
     conn = pyodbc.connect(conn_str)
