@@ -2,21 +2,21 @@ import { useEffect, useState } from "react";
 
 export default function EventsPage() {
   const [favorites, setFavorites] = useState([]);
-  const userId = '0738';
+  const [participatedEvents, setParticipatedEvents] = useState([]);
+  // idを直接定義
+  const id = '0738';
 
+  // toggleFavorite関数を定義
   function toggleFavorite(eventId) {
-    const userId = '0738';
     setFavorites(prev => {
-      // すでにお気に入りの場合は何もしない（色も維持）
       if (prev.includes(eventId)) {
         return prev;
       }
       const updated = [...prev, eventId];
-      // お気に入り追加時のみAPI呼び出し
       fetch("https://0x0-showevent-hbbadxcxh9a4bzhu.japaneast-01.azurewebsites.net/api/favorite?code=zsOO_WgPGY9dtEN_tkki1bHWPy8XYJQoQPo2G7ONmvsoAzFusJrTJg%3D%3D", {
         method: "POST",
         headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ event_id: eventId, id: userId })
+        body: JSON.stringify({ event_id: eventId, id }) // ← userIdではなくidに統一
       })
       .then(async res => {
         if (!res.ok) {
@@ -37,11 +37,9 @@ export default function EventsPage() {
         return res.text();
       })
       .then(data => {
-        // 登録成功時に通知
         alert('お気に入りに登録しました');
       })
       .catch(err => {
-        // 既にalert済みなのでconsoleのみ
         console.error("お気に入り登録APIエラー", err);
       });
       localStorage.setItem("favorites", JSON.stringify(updated));
@@ -60,18 +58,17 @@ export default function EventsPage() {
 
   useEffect(() => {
     // お気に入り情報をAPIから取得
-    fetch(`https://0x0-showevent-hbbadxcxh9a4bzhu.japaneast-01.azurewebsites.net/api/favorites?user_id=${userId}`)
+    fetch(`https://0x0-showevent-hbbadxcxh9a4bzhu.japaneast-01.azurewebsites.net/api/favorites?user_id=${id}&code=zsOO_WgPGY9dtEN_tkki1bHWPy8XYJQoQPo2G7ONmvsoAzFusJrTJg%3D%3D`)
       .then(res => res.json())
       .then(data => {
-        // dataはevent_idの配列を想定
         setFavorites(Array.isArray(data) ? data : []);
       })
       .catch(() => setFavorites([]));
 
-    fetch("/api/event/list")
+    // イベント一覧取得API
+    fetch("https://0x0-showevent-hbbadxcxh9a4bzhu.japaneast-01.azurewebsites.net/api/showEvent?code=KjUCLx4igb6FiJ3ZtQKowVUUk9MgUtPSuBhPrMam2RwxAzFuTt1T_w%3D%3D")
       .then((res) => res.json())
       .then((data) => {
-        // is_draftが1のものは除外 & 締切期限が過ぎたものは除外
         const now = new Date();
         const filtered = Array.isArray(data)
           ? data.filter(event => {
@@ -87,7 +84,7 @@ export default function EventsPage() {
         setError("データ取得エラー: " + err.message);
       });
 
-    // カテゴリー一覧をCATEGORYSテーブルAPIから取得
+    // カテゴリー一覧取得API
     fetch("https://0x0-showevent-hbbadxcxh9a4bzhu.japaneast-01.azurewebsites.net/api/categories?code=qPu7q4iQBMrEMTPaYXSYNOrzTnAm5yplhzIJ9JfIq-vWAzFukZ5pSA%3D%3D")
       .then(res => res.json())
       .then(data => {
@@ -96,11 +93,18 @@ export default function EventsPage() {
       .catch(err => {
         setError("カテゴリー取得エラー: " + err.message);
       });
+
+    // 参加済みイベント取得API
+    fetch("https://0x0-participation-d7fqb7h3dpcqcxek.japaneast-01.azurewebsites.net/api/event/participate?code=IqAEzEm_tdgsaLYblJjNZChDOjX7TKk2FDdM9zV2yMqFAzFufBImGw%3D%3D&user_id=" + id)
+      .then(res => res.json())
+      .then(data => {
+        setParticipatedEvents(Array.isArray(data) ? data : []);
+      })
+      .catch(() => setParticipatedEvents([]));
   }, []);
 
   // 表示するカラムを限定
   const filteredKeys = ["event_title", "event_datetime", "deadline", "location"];
-  const displayKeys = ["event_title", "event_datetime", "deadline", "location", "participants_status"];
 
   return (
     <div>
@@ -216,7 +220,7 @@ export default function EventsPage() {
                   ))}
                   <td>{`${event.current_participants ?? 0}/${event.max_participants ?? 0}`}</td>
                   <td style={{ textAlign: 'center' }}>
-                    <button onClick={() => window.location.href = `/event-detail?event_id=${event.event_id}`}>詳細</button>
+                    <button onClick={() => window.location.href = `/event/detail?event_id=${event.event_id}&id=${id}`}>詳細</button>
                   </td>
                   <td style={{ textAlign: 'center' }}>
                     <button
@@ -241,3 +245,4 @@ export default function EventsPage() {
     </div>
   );
 }
+
